% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-dictionaries_functions.R
\name{as_data_dict_shape}
\alias{as_data_dict_shape}
\title{xxx xxx xxx}
\usage{
as_data_dict_shape(object)
}
\arguments{
\item{object}{xxx xxx xxx}
}
\value{
xxx xxx xxx
}
\description{
xxx xxx xxx
}
\details{
xxx xxx xxx
}
\examples{
\dontrun{

# example

}

}
