% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-catalogue_functions.R
\name{as_dataset}
\alias{as_dataset}
\title{xxx xxx xxx}
\usage{
as_dataset(object, col_id = NULL)
}
\arguments{
\item{object}{xxx xxx xxx}

\item{col_id}{xxx xxx xxx}
}
\value{
xxx xxx xxx
}
\description{
xxx xxx xxx
}
\details{
xxx xxx xxx
}
\examples{
\dontrun{

# example

}

}
