% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-dictionaries_functions.R
\name{as_mlstr_data_dict}
\alias{as_mlstr_data_dict}
\title{xxx xxx xxx}
\usage{
as_mlstr_data_dict(object, as_data_dict = FALSE)
}
\arguments{
\item{object}{xxx xxx xxx}

\item{as_data_dict}{xxx xxx xxx}
}
\value{
xxx xxx xxx.
}
\description{
xxx xxx xxx.
}
\examples{
\dontrun{
# Example 1: xxx xxx xxx.

}

}
