% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-summary_functions.R
\name{check_cat_values}
\alias{check_cat_values}
\title{xxx xxx xxx}
\usage{
check_cat_values(dataset, data_dict)
}
\arguments{
\item{dataset}{xxx xxx xxx}

\item{data_dict}{xxx xxx xxx}
}
\value{
xxx xxx xxx.
}
\description{
xxx xxx xxx.
}
\examples{
\dontrun{

# Example 1: xxx xxx xxx.


}

}
