% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-data_QA.R
\name{check_data_dict_variables}
\alias{check_data_dict_variables}
\title{xxx xxx xxx}
\usage{
check_data_dict_variables(data_dict)
}
\arguments{
\item{data_dict}{xxx xxx xxx}
}
\value{
xxx xxx xxx.
}
\description{
xxx xxx xxx.
}
\examples{
\dontrun{

# Example 1: xxx xxx xxx.


}

}
