% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-data_QA.R
\name{check_dataset_variables}
\alias{check_dataset_variables}
\title{xxx xxx xxx}
\usage{
check_dataset_variables(data, data_dict)
}
\arguments{
\item{data}{xxx xxx xxx}

\item{data_dict}{xxx xxx xxx}
}
\value{
xxx xxx xxx.
}
\description{
xxx xxx xxx.
}
\examples{
\dontrun{

# Example 1: xxx xxx xxx.


}

}
