% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-summary_functions.R
\name{check_existing_variable}
\alias{check_existing_variable}
\title{xxx xxx xxx}
\usage{
check_existing_variable(data_dict, dataset)
}
\arguments{
\item{data_dict}{xxx xxx xxx}

\item{dataset}{xxx xxx xxx}
}
\value{
xxx xxx xxx.
}
\description{
xxx xxx xxx.
}
\examples{
\dontrun{

# Example 1: xxx xxx xxx.


}

}
