% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-data_QA.R
\name{check_name_standards}
\alias{check_name_standards}
\title{xxx xxx xxx}
\usage{
check_name_standards(var_names)
}
\arguments{
\item{var_names}{xxx xxx xxx}
}
\value{
xxx xxx xxx.
}
\description{
xxx xxx xxx.
}
\examples{
\dontrun{

# Example 1: xxx xxx xxx.


}

}
