% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-dictionaries_functions.R
\name{data_dict_expand}
\alias{data_dict_expand}
\title{xxx xxx xxx}
\usage{
data_dict_expand(
  data_dict,
  from = "Variables",
  name_prefix = "Categories::",
  to = "Categories"
)
}
\arguments{
\item{data_dict}{xxx xxx xxx}

\item{from}{xxx xxx xxx}

\item{name_prefix}{xxx xxx xxx}

\item{to}{xxx xxx xxx}
}
\value{
xxx xxx xxx.
}
\description{
xxx xxx xxx.
}
\examples{
\dontrun{
# Example 1: xxx xxx xxx.


}

}
