% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-dictionaries_functions.R
\name{data_dict_extract}
\alias{data_dict_extract}
\title{Create a data dictionary in the Maelstrom Research format from a dataset}
\usage{
data_dict_extract(data)
}
\arguments{
\item{data}{xxx xxx xxx.}
}
\value{
A list of two tibbles: Variables and Categories. Variables has all the
variable names extracted from the dataset column names as rows and the Maelstrom
Research descriptors standards of variables as columns. Categories has all the category
names extracted from the “categories” argument as rows and the Maelstrom Research
descriptors standard of categories as columns.
}
\description{
Creates a data dictionary in the Maelstrom Research formats (with "Variables" and "Categories"
in separate tibbles and standard columns in each) from any dataset in tibble format.
If the input dataset has no associated metadata, a data dictionary with a minimal
of information is created from the column (variable) names to create the data
dictionary structure required for harmonizR (all columns except variable names
will be blank).
}
\details{
Must provide an input data dictionary in tibble format, with or without metadata
in addition to column names.
}
\examples{
\dontrun{
# use case 1: create a data dictionary from any dataset
data_dict_extract(iris)

# use case 2: create a data dictionary with project and categorical variables
}

}
