% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-dictionaries_functions.R
\name{data_dict_filter}
\alias{data_dict_filter}
\title{xxx xxx xxx}
\usage{
data_dict_filter(
  data_dict,
  filter_var = NULL,
  filter_cat = NULL,
  filter_all = NULL
)
}
\arguments{
\item{data_dict}{xxx xxx xxx}

\item{filter_var}{xxx xxx xxx}

\item{filter_cat}{xxx xxx xxx}

\item{filter_all}{xxx xxx xxx}
}
\value{
xxx xxx xxx.
}
\description{
xxx xxx xxx.
}
\examples{
\dontrun{
# Example 1: xxx xxx xxx.

}

}
