% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-dictionaries_functions.R
\name{data_dict_flatten}
\alias{data_dict_flatten}
\title{xxx xxx xxx}
\usage{
data_dict_flatten(
  data_dict,
  from = "Categories",
  to = "Variables",
  name_prefix = "Categories::"
)
}
\arguments{
\item{data_dict}{xxx xxx xxx}

\item{from}{xxx xxx xxx}

\item{to}{xxx xxx xxx}

\item{name_prefix}{xxx xxx xxx}
}
\value{
xxx xxx xxx.
}
\description{
xxx xxx xxx.
}
\examples{
\dontrun{
# Example 1: xxx xxx xxx.

}

}
