% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-dictionaries_functions.R
\name{data_dict_group_split}
\alias{data_dict_group_split}
\title{xxx xxx xxx}
\usage{
data_dict_group_split(data_dict, ...)
}
\arguments{
\item{data_dict}{xxx xxx xxx}

\item{...}{xxx xxx xxx}
}
\value{
xxx xxx xxx.
}
\description{
xxx xxx xxx.
}
\examples{
\dontrun{
# Example 1: xxx xxx xxx.

}

}
