% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-dictionaries_functions.R
\name{data_dict_match_dataset}
\alias{data_dict_match_dataset}
\title{xxx xxx xxx}
\usage{
data_dict_match_dataset(
  data,
  data_dict,
  data_dict_apply = FALSE,
  output = c("data", "data_dict")
)
}
\arguments{
\item{data}{xxx xxx xxx}

\item{data_dict}{xxx xxx xxx}

\item{data_dict_apply}{xxx xxx xxx}

\item{output}{xxx xxx xxx}
}
\value{
xxx xxx xxx.
}
\description{
xxx xxx xxx.
}
\examples{
\dontrun{

# Example 1: xxx xxx xxx.


}

}
