% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-dictionaries_functions.R
\name{data_dict_pivot_wider}
\alias{data_dict_pivot_wider}
\title{xxx xxx xxx}
\usage{
data_dict_pivot_wider(data_dict, taxonomy = NULL)
}
\arguments{
\item{data_dict}{xxx xxx xxx}

\item{taxonomy}{xxx xxx xxx}
}
\value{
xxx xxx xxx.
}
\description{
xxx xxx xxx.
}
\examples{
\dontrun{
# Example 1: xxx xxx xxx.

}

}
