% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-catalogue_functions.R
\name{data_extract}
\alias{data_extract}
\title{xxx xxx xxx}
\usage{
data_extract(data_dict, data_dict_apply = FALSE)
}
\arguments{
\item{data_dict}{xxx xxx xxx}

\item{data_dict_apply}{xxx xxx xxx}
}
\value{
xxx xxx xxx.
}
\description{
xxx xxx xxx.
}
\examples{
\dontrun{

# Example 1: xxx xxx xxx.


}

}
