% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-dictionaries_functions.R
\name{is_mlstr_data_dict}
\alias{is_mlstr_data_dict}
\title{xxx xxx xxx}
\usage{
is_mlstr_data_dict(object)
}
\arguments{
\item{object}{xxx xxx xxx}
}
\value{
xxx xxx xxx
}
\description{
xxx xxx xxx
}
\details{
xxx xxx xxx
}
\examples{
\dontrun{

# example

}

}
