% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-catalogue_functions.R
\name{is_taxonomy}
\alias{is_taxonomy}
\title{xxx xxx xxx}
\usage{
is_taxonomy(object)
}
\arguments{
\item{object}{xxx xxx xxx}
}
\value{
xxx xxx xxx
}
\description{
xxx xxx xxx
}
\details{
xxx xxx xxx
}
\examples{
\dontrun{

# example

}

}
