% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-visualize.R
\name{open_visual_report}
\alias{open_visual_report}
\title{xxx xxx xxx}
\usage{
open_visual_report(report_name)
}
\arguments{
\item{report_name}{xxx xxx xxx}
}
\value{
xxx xxx xxx.
}
\description{
xxx xxx xxx.
}
\examples{
\dontrun{

# Example 1: xxx xxx xxx.


}

}
