% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-catalogue_functions.R
\name{study_create}
\alias{study_create}
\title{xxx xxx xxx}
\usage{
study_create(dataset_list, data_dict_apply = FALSE)
}
\arguments{
\item{dataset_list}{xxx xxx xxx}

\item{data_dict_apply}{xxx xxx xxx}
}
\value{
xxx xxx xxx.
}
\description{
xxx xxx xxx.
}
\examples{
\dontrun{

# Example 1: xxx xxx xxx.


}

}
