% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-summary_functions.R
\name{summary_variables_categorical}
\alias{summary_variables_categorical}
\title{xxx xxx xxx}
\usage{
summary_variables_categorical(name_var, dict_var)
}
\arguments{
\item{name_var}{xxx xxx xxx}

\item{dict_var}{xxx xxx xxx}
}
\value{
xxx xxx xxx.
}
\description{
xxx xxx xxx.
}
\examples{
\dontrun{

# Example 1: xxx xxx xxx.


}

}
