% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-catalogue_functions.R
\name{valueType_adjust}
\alias{valueType_adjust}
\title{xxx xxx xxx}
\usage{
valueType_adjust(from, to = NULL)
}
\arguments{
\item{from}{xxx xxx xxx}

\item{to}{xxx xxx xxx}
}
\value{
xxx xxx xxx.
}
\description{
xxx xxx xxx.
}
\examples{
\dontrun{

# Example 1: xxx xxx xxx.


}

}
