% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-catalogue_functions.R
\name{valueType_guess}
\alias{valueType_guess}
\title{Apply all the valueTypes of a data dictionary to a dataset}
\usage{
valueType_guess(x)
}
\arguments{
\item{x}{A xxx xxx}
}
\value{
The dataset with the data dictionary valueType
applied to each variable
}
\description{
xxx xxx xxx
}
\details{
Must provide an input dataset in tibble format
}
\examples{
\dontrun{
# use case 1: Apply valueType without specifying a data dictionary
as_valueType("1") \%>\% typeof()

tibble(iris \%>\% mutate(Species = as.character(Species))) \%>\%
  mutate(across(c(starts_with("Sepal")), ~ as_valueType(.,"integer")))
}

}
