% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-catalogue_functions.R
\name{valueType_of}
\alias{valueType_of}
\title{xxx xxx xxx}
\usage{
valueType_of(x)
}
\arguments{
\item{x}{xxx xxx xxx}
}
\value{
xxx xxx xxx.
}
\description{
xxx xxx xxx.
}
\examples{
\dontrun{

# Example 1: xxx xxx xxx.


}

}
