% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-catalogue_functions.R
\name{valueType_self_adjust}
\alias{valueType_self_adjust}
\title{xxx xxx xxx}
\usage{
valueType_self_adjust(...)
}
\arguments{
\item{...}{xxx xxx xxx}
}
\value{
xxx xxx xxx.
}
\description{
xxx xxx xxx.
}
\examples{
\dontrun{

# Example 1: xxx xxx xxx.


}

}
